requirements.txt
Django==5.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pillow==10.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.7.0
django-filter==25.1
SpeechRecognition==3.14.2
pydub==0.25.1
ffmpeg-python==0.2.0
#
autoflake==2.3.1
black==24.8.0
pre-commit==3.8.0
click==8.1.7
isort==5.13.2
flake8==7.1.1
flake8-formatter-junit-xml==0.0.6